Hitboxes and models

+++ Project MOMENTUM/Assets/Scripts/Interfaces/IDamagable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IDamagable
 {
-    void Damage(GameObject attacker, int damage);
+    void Damage(GameObject attacker, float damage);
 }
